Turn Prompts tab into Modes tab and move support prompts to Settings
