+++ src/package.nls.ca.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "Servidors MCP",
-	"command.prompts.title": "Indicacions",
+	"command.prompts.title": "Modes",
 	"command.history.title": "Historial",
 	"command.openInEditor.title": "Obrir a l'Editor",
 	"command.settings.title": "Configuració",
