+++ src/package.nls.de.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "MCP Server",
-	"command.prompts.title": "Prompts",
+	"command.prompts.title": "Modi",
 	"command.history.title": "Verlauf",
 	"command.openInEditor.title": "Im Editor Öffnen",
 	"command.settings.title": "Einstellungen",
