+++ src/package.nls.es.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "Servidores MCP",
-	"command.prompts.title": "Indicaciones",
+	"command.prompts.title": "Modos",
 	"command.history.title": "Historial",
 	"command.openInEditor.title": "Abrir en Editor",
 	"command.settings.title": "Configuración",
