+++ src/package.nls.fr.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "Serveurs MCP",
-	"command.prompts.title": "Invites",
+	"command.prompts.title": "Modes",
 	"command.history.title": "Historique",
 	"command.openInEditor.title": "Ouvrir dans l'Éditeur",
 	"command.settings.title": "Paramètres",
