+++ src/package.nls.hi.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "एमसीपी सर्वर",
-	"command.prompts.title": "प्रॉम्प्ट्स",
+	"command.prompts.title": "मोड्स",
 	"command.history.title": "इतिहास",
 	"command.openInEditor.title": "एडिटर में खोलें",
 	"command.settings.title": "सेटिंग्स",
