+++ src/package.nls.it.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "Server MCP",
-	"command.prompts.title": "Prompt",
+	"command.prompts.title": "Modi",
 	"command.history.title": "Cronologia",
 	"command.openInEditor.title": "Apri nell'Editor",
 	"command.settings.title": "Impostazioni",
