+++ src/package.nls.ja.json
@@ -6,7 +6,7 @@
 	"views.activitybar.title": "Roo Code",
 	"command.newTask.title": "新しいタスク",
 	"command.mcpServers.title": "MCPサーバー",
-	"command.prompts.title": "プロンプト",
+	"command.prompts.title": "モード",
 	"command.history.title": "履歴",
 	"command.openInEditor.title": "エディタで開く",
 	"command.settings.title": "設定",
