+++ src/package.nls.json
@@ -6,7 +6,7 @@
 	"views.activitybar.title": "Roo Code",
 	"command.newTask.title": "New Task",
 	"command.mcpServers.title": "MCP Servers",
-	"command.prompts.title": "Prompts",
+	"command.prompts.title": "Modes",
 	"command.history.title": "History",
 	"command.openInEditor.title": "Open in Editor",
 	"command.settings.title": "Settings",
