+++ src/package.nls.nl.json
@@ -6,7 +6,7 @@
 	"views.activitybar.title": "Roo Code",
 	"command.newTask.title": "Nieuwe Taak",
 	"command.mcpServers.title": "MCP Servers",
-	"command.prompts.title": "Prompts",
+	"command.prompts.title": "Modi",
 	"command.history.title": "Geschiedenis",
 	"command.openInEditor.title": "Openen in Editor",
 	"command.settings.title": "Instellingen",
