+++ src/package.nls.pl.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "Serwery MCP",
-	"command.prompts.title": "Podpowiedzi",
+	"command.prompts.title": "Tryby",
 	"command.history.title": "Historia",
 	"command.openInEditor.title": "Otwórz w Edytorze",
 	"command.settings.title": "Ustawienia",
