+++ src/package.nls.pt-BR.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "Servidores MCP",
-	"command.prompts.title": "Prompts",
+	"command.prompts.title": "Modos",
 	"command.history.title": "Histórico",
 	"command.openInEditor.title": "Abrir no Editor",
 	"command.settings.title": "Configurações",
