+++ src/package.nls.ru.json
@@ -6,7 +6,7 @@
 	"views.activitybar.title": "Roo Code",
 	"command.newTask.title": "Новая задача",
 	"command.mcpServers.title": "MCP серверы",
-	"command.prompts.title": "Промпты",
+	"command.prompts.title": "Режимы",
 	"command.history.title": "История",
 	"command.openInEditor.title": "Открыть в редакторе",
 	"command.settings.title": "Настройки",
