+++ src/package.nls.tr.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "MCP Sunucuları",
-	"command.prompts.title": "Komut İstemleri",
+	"command.prompts.title": "Modlar",
 	"command.history.title": "Geçmiş",
 	"command.openInEditor.title": "Düzenleyicide Aç",
 	"command.settings.title": "Ayarlar",
