+++ src/package.nls.vi.json
@@ -17,7 +17,7 @@
 	"views.contextMenu.label": "Roo Code",
 	"views.terminalMenu.label": "Roo Code",
 	"command.mcpServers.title": "Máy Chủ MCP",
-	"command.prompts.title": "Lời Nhắc",
+	"command.prompts.title": "Chế Độ",
 	"command.history.title": "Lịch Sử",
 	"command.openInEditor.title": "Mở trong Trình Soạn Thảo",
 	"command.settings.title": "Cài Đặt",
