+++ webview-ui/src/i18n/locales/hi/prompts.json
@@ -1,5 +1,5 @@
 {
-	"title": "प्रॉम्प्ट्स",
+	"title": "मोड्स",
 	"done": "हो गया",
 	"modes": {
 		"title": "मोड्स",
@@ -64,7 +64,8 @@
 			"apiConfigDescription": "आप प्रॉम्प्ट्स को बढ़ाने के लिए हमेशा उपयोग करने के लिए एक API कॉन्फ़िगरेशन चुन सकते हैं, या बस वर्तमान में चयनित का उपयोग कर सकते हैं",
 			"useCurrentConfig": "वर्तमान में चयनित API कॉन्फ़िगरेशन का उपयोग करें",
 			"testPromptPlaceholder": "वृद्धि का परीक्षण करने के लिए एक प्रॉम्प्ट दर्ज करें",
-			"previewButton": "प्रॉम्प्ट वृद्धि का पूर्वावलोकन"
+			"previewButton": "प्रॉम्प्ट वृद्धि का पूर्वावलोकन",
+			"testEnhancement": "वृद्धि का परीक्षण करें"
 		},
 		"types": {
 			"ENHANCE": {
