+++ webview-ui/src/i18n/locales/hi/settings.json
@@ -28,9 +28,13 @@
 		"notifications": "सूचनाएँ",
 		"contextManagement": "संदर्भ",
 		"terminal": "टर्मिनल",
+		"prompts": "प्रॉम्प्ट्स",
 		"experimental": "प्रायोगिक",
 		"language": "भाषा",
 		"about": "परिचय"
+	},
+	"prompts": {
+		"description": "प्रॉम्प्ट्स को बेहतर बनाना, कोड की व्याख्या करना और समस्याओं को ठीक करना जैसी त्वरित कार्रवाइयों के लिए उपयोग किए जाने वाले सहायक प्रॉम्प्ट्स को कॉन्फ़िगर करें। ये प्रॉम्प्ट्स Roo को सामान्य विकास कार्यों के लिए बेहतर सहायता प्रदान करने में मदद करते हैं।"
 	},
 	"codeIndex": {
 		"title": "कोडबेस इंडेक्सिंग",
