+++ webview-ui/src/i18n/locales/ko/settings.json
@@ -28,9 +28,13 @@
 		"notifications": "알림",
 		"contextManagement": "컨텍스트",
 		"terminal": "터미널",
+		"prompts": "프롬프트",
 		"experimental": "실험적",
 		"language": "언어",
 		"about": "Roo Code 정보"
+	},
+	"prompts": {
+		"description": "프롬프트 향상, 코드 설명, 문제 해결과 같은 빠른 작업에 사용되는 지원 프롬프트를 구성합니다. 이러한 프롬프트는 Roo가 일반적인 개발 작업에 대해 더 나은 지원을 제공하는 데 도움이 됩니다."
 	},
 	"codeIndex": {
 		"title": "코드베이스 인덱싱",
