+++ webview-ui/src/i18n/locales/ru/prompts.json
@@ -1,5 +1,5 @@
 {
-	"title": "Промпты",
+	"title": "Режимы",
 	"done": "Готово",
 	"modes": {
 		"title": "Режимы",
@@ -64,7 +64,8 @@
 			"apiConfigDescription": "Вы можете выбрать конфигурацию API, которая всегда будет использоваться для улучшения промптов, или использовать текущую выбранную",
 			"useCurrentConfig": "Использовать текущую конфигурацию API",
 			"testPromptPlaceholder": "Введите промпт для тестирования улучшения",
-			"previewButton": "Просмотреть улучшенный промпт"
+			"previewButton": "Просмотреть улучшенный промпт",
+			"testEnhancement": "Тестировать улучшение"
 		},
 		"types": {
 			"ENHANCE": {
