+++ webview-ui/src/i18n/locales/ru/settings.json
@@ -28,9 +28,13 @@
 		"notifications": "Уведомления",
 		"contextManagement": "Контекст",
 		"terminal": "Терминал",
+		"prompts": "Промпты",
 		"experimental": "Экспериментальное",
 		"language": "Язык",
 		"about": "О Roo Code"
+	},
+	"prompts": {
+		"description": "Настройте промпты поддержки, используемые для быстрых действий, таких как улучшение промптов, объяснение кода и исправление проблем. Эти промпты помогают Roo обеспечить лучшую поддержку для общих задач разработки."
 	},
 	"codeIndex": {
 		"title": "Индексация кодовой базы",
